docs: revamp README.md description wording and links
The previous README.md content was improved for better description and understanding. Links are updated to the most relevant ones, helping users to understand the purpose of the project more clearly. Additionally, the instructions on how to initiate new projects using Slim framework are made more clear, benefiting both novice and seasoned developers.