ci: replace inline release creation with a reusable action
To simplify the process of creating a new release draft in GitHub actions, I replaced the inline script for the 'Create Release Draft' step in 'build.yml' with 'nekofar/create-github-release@v1.0.0', a custom reusable action. This not only makes our workflow easier to read and maintain but also standardizes the way we create release drafts across different projects. The pre-existing environment variable 'GITHUB_TOKEN' is not needed anymore as the reusable action manages it internally.